BUG#: 6103 TITLE: Use Uint32 instead of size_t in List template DESCRIPTION: Use Uint32 in the List template to remove the need for explicit type conversions from size_t to Uint32.

+++ src/Pegasus/Common/Cimom.cpp
@@ -850,7 +850,7 @@ Boolean cimom::moduleChange(struct timeval last)
 
 Uint32 cimom::getModuleCount()
 {
-    return (Uint32)_modules.size();
+    return _modules.size();
 }
 
 Uint32 cimom::getModuleIDs(Uint32* ids, Uint32 count)
@@ -877,7 +877,7 @@ Uint32 cimom::getModuleIDs(Uint32* ids, Uint32 count)
         count--;
     }
 
-    return (Uint32)_modules.size();
+    return _modules.size();
 }
 
 AsyncOpNode* cimom::get_cached_op()
